fix for game parsing in non-ladder mode

+++ src/utility/GameParser.java
@@ -51,16 +51,22 @@ public class GameParser
 				return;
 			}
 			
-			FileUtils.lockFile(ServerSettings.Instance().GamesListFile + ".lock", 5, 100, 60000);
-		
+			if (ServerSettings.Instance().LadderMode)
+			{
+				FileUtils.lockFile(ServerSettings.Instance().GamesListFile + ".lock", 5, 100, 60000);
+			}
+			
 			BufferedReader br = new BufferedReader(new FileReader(ServerSettings.Instance().GamesListFile));
 			parseGames(br);
 			br.close();
 			
 			// in ladder mode games list should be deleted after parsing
-			FileUtils.DeleteFile(gamesFile);
+			if (ServerSettings.Instance().LadderMode)
+			{
+				FileUtils.DeleteFile(gamesFile);
+				FileUtils.unlockFile(ServerSettings.Instance().GamesListFile + ".lock");
+			}
 			
-			FileUtils.unlockFile(ServerSettings.Instance().GamesListFile + ".lock");
 		}
 		catch (FileNotFoundException e) 
 		{
